File and directory were updated on different files.
