+++ src/com/jalasoft/search_files/search/asset.py
@@ -1,6 +1,7 @@
 import os
 import datetime
 
+
 class Asset(object):
 
     def __init__(self, path, is_file):
@@ -40,28 +41,3 @@ class Asset(object):
         return self._is_file
 
 
-class File(Asset):
-
-    def __init__(self, path, file_name, is_file):
-        super().__init__(path, is_file)
-
-        self._file_name, self._separator, self._extension = file_name.partition(".")
-
-    def get_file_name(self):
-        return self._file_name
-
-    def get_extension(self):
-        return self._extension
-
-    def get_separator(self):
-        return self._separator
-
-class Directory(Asset):
-
-    def __init__(self, path, dir_name, is_file):
-        super().__init__(path, is_file)
-        self.count_content = 0
-        self.dir_name = dir_name
-
-    def get_dir_name(self):
-        return self.dir_name
