+++ src/com/jalasoft/search_files/search/search.py
@@ -229,7 +229,7 @@ class Search(object):
             if isinstance(item, File):
                 item_mb = item.get_size() / (1024 * 1024)
                 if less_size != 0 and bigger_size != 0:
-                    if item_mb <= less_size and item_mb >= bigger_size:
+                    if less_size >= item_mb >= bigger_size:
                         result_asert_list.append(item)
                 elif less_size != 0 and bigger_size == 0:
                     if item_mb <= less_size :
